클라이언트 업데이트

+++ 최규화/Mod_Client/프레임워크 3.4/CGameObject.h
@@ -141,7 +141,7 @@ public:
 	bool isShieldOn = false;
 	bool isHit1 = false;
 
-	float TexStride = 0.f;
+	int TexStride = 0;
 
 	//벽들에 굳이 마우스를 움직일때마다 체크할 필요는 없으므로 추가함. 또 벽은 또 벽대로 뭔가 처리할게 있을것같음.
 	Obj_State obs = Dynamic;
